LPS-112275 Use mock inside EditTableView unit test
